+++ packages/BilibiliRecorder/src/index.ts
@@ -256,6 +256,7 @@ const checkLiveStatusAndRecord: Recorder["checkLiveStatusAndRecord"] = async fun
       formatName: streamOptions.format_name as "flv" | "ts" | "fmp4",
       disableDanma: this.disableProvideCommentsWhenRecording,
       videoFormat: this.videoFormat,
+      debugLevel: this.debugLevel ?? "none",
     },
     onEnd,
     async () => {