+++ packages/DouYuRecorder/src/index.ts
@@ -236,6 +236,7 @@ const checkLiveStatusAndRecord: Recorder["checkLiveStatusAndRecord"] = async fun
         getSavePath({ owner, title: opts.title ?? title, startTime: opts.startTime }),
       disableDanma: this.disableProvideCommentsWhenRecording,
       videoFormat: this.videoFormat ?? "auto",
+      debugLevel: this.debugLevel ?? "none",
     },
     onEnd,
     async () => {