+++ packages/liveManager/src/recorder.ts
@@ -71,6 +71,9 @@ export interface RecorderCreateOpts<E extends AnyObject = UnknownObject> {
   useServerTimestamp?: boolean;
   // 可持久化的额外字段，让 provider、manager 开发者可以有更多 customize 的空间
   extra?: Partial<E>;
+  /** 调试等级 */
+  debugLevel?: "none" | "basic" | "verbose";
+  /** 缓存 */
   cache: Cache;
 }