+++ packages/HuYaRecorder/CHANGELOG.md
@@ -1,5 +1,6 @@
-# 1.8.0
+# Next
 
+- 新增`debugLevel`参数，支持`none`、`basic`、`verbose`
 - 支持标题黑名单
 - 触发标题黑名单设定额外状态